+++ Quicksilver/Code-QuickStepCore/QSCommand.m
@@ -401,7 +401,7 @@ NSTimeInterval QSTimeIntervalForString(NSString *intervalString) {
 	if (VERBOSE) NSLog(@"Execute Command: %@", self);
 #endif
 	QSInterfaceController *controller = [(QSController *)[NSApp delegate] interfaceController];
-	int argumentCount = [(QSAction *)actionObject argumentCount];
+	NSInteger argumentCount = [(QSAction *)actionObject argumentCount];
 	if (argumentCount == 2 && ![actionObject indirectOptional] && (!indirectObject || [[indirectObject primaryType] isEqualToString:QSTextProxyType])) {
 		// indirect object required, but is either missing or asking for text input
 		if (!indirectObject) {
